# [RFC] Speed up Axes.bar