@jklymak 2021-04-27
Thanks!  I think this idea has been discussed before.  I'm not sure of the technical issues, but in general it seems reasonable.  

Some of your image changes are quite large according to the tests so it would be good to see before and after and decide what is wrong.  The way to do that is to check in the new images as part of the PR (in tests/baseline_images) and we can see the difference.