@ivirshup 2021-04-27
@jklymak thanks for the quick look! I've added a TODO to the top of the PR with the problems I've noticed in the test plots.

Currently looking at legends for `hist`. Not totally sure how to approach it, but am currently thinking of modifying this part of `Axes.hist`

```python
        # If None, make all labels None (via zip_longest below); otherwise,
        # cast each element to str, but keep a single str as it.
        labels = [] if label is None else np.atleast_1d(np.asarray(label, str))
        for patch, lbl in itertools.zip_longest(patches, labels):
            if patch:
                p = patch[0]
                p.update(kwargs)
                if lbl is not None:
                    p.set_label(lbl)
```

And instead adding a new something (`Artist` I guess?) to the plot from the patches that would have been assigned labels.

<details>
<summary> example </summary>

```python
        labels = [] if label is None else np.atleast_1d(np.asarray(label, str))
        legend_patches = []
        for patch, lbl in itertools.zip_longest(patches, labels):
            if patch:
                p = patch[0]
                p.update(kwargs)
                if lbl is not None:
                    p.set_label(lbl)
                    legend_patches.append(p)
                for p in patch[1:]:
                    p.update(kwargs)
                    p.set_label('_nolegend_')

        self.add_somehow(legend_patches)
```

</details>

Is there a more elegant way to do this?

-------------------

@mwaskom, is it possible to expand the collections into lists of patches? Maybe a helper around this?