@tacaswell 2021-04-28
Changing the return type is definitively an API change.  Any user who is trying to do something with the returned is going to have their code broken.

You might be able to sub-class the Collection and make it look enough like the current return type, but it will be a lot of work as things like

```
bars = ax.bar(....)
bars[1].set_color('green')
bars[2].set_visible(False)
bars[3].set_linewidth(15)
*_, last, two = bars
```

all should keep working.

-----

If your problem is primarily histograms @ivirshup you may do better either passing `histtype='step'` or using the new in mpl34 `ax.stairs` method.  If you really want the bar look on histograms adding a new histtype (and accepting a bit more opt-in type instability in `hist`) may be easier to get done than changing the return type of all `bar` calls.