@jklymak 2021-04-28
maybe we could have `ax.bar(...,  quick=True)`?  Though `ax.histbar` isn't a bad idea.