@ivirshup 2021-04-28
@tacaswell, my thing was mainly that I didn't want to have to consider the size of my data and choose the keyword arguments. Particularly as a library author, having changes to the plots around arbitrary data size cutoffs is not going to be fun to maintain. Also I think people generally are expecting something that looks like bars when the bins are wide enough to be able to tell.

I think your point is totally fair. I wasn't really sure what was considered to be key parts of the API though, since arguably any changes to public attributes of the object is breaking. I wasn't planning on changing the return type, just it arguably wouldn't be a particularly useful value. Basically, my thinking was if changing the contents of `ax.containers` isn't breaking, is changing the behaviour of `res = ax.bar(...); res[0].set_facecolor(...)` breaking? But also, if the first is breaking, it would be difficult to make any improvements to this library.

TBH, if this got rejected here, my plan was to see if @mwaskom would be up for replacing the use of `ax.bar` in `seaborn` with some `Collection`s, since initial tests make it look like this will cut plot time for large histograms by at least 50%.

--------------------------

From looking at the code a bit, I was actually wondering if the plan was to deprecate the `containers` module, since it's only used in a small subset of the library (specifically for bar and stem charts, I think). A little confusingly, some things that are annotated as `Container` actually expect subclasses of `ToolContainerBase` which is unrelated.

https://github.com/matplotlib/matplotlib/blob/a94acb3d2c6829d7ce3a4ef2cc33058c677e0526/lib/matplotlib/backend_tools.py#L1008-L1009

My point here being, maybe this would be worth it for 4.0? It can be quite slow at the moment, and the machinery for making it fast is all here. Using a Collection instead of a Container would also bring this function more in line with the rest of package.