@timhoffm 2021-04-28
Every change that can break or change the result of currently valid user code is a breaking change to us. 

> if the first is breaking, it would be difficult to make any improvements to this library.

Unfortunately, it is. This is the downside of a very large user base. We have to be extermely careful with these. 

Using a Collection for the bars is reasonable. Probably the way forward is similar to `hist(..., histtype=...)`: Add a parameter to make the Artist configurable. The default stays on isolated bars, using a collection is opt-in. The default can be switched later.


A clever in between way may be returning a proxy in  `container.patch`, that will resolve the PatchCollection into individual rectangles when trying to access individual elements. If the user does not mess with individual rectangles, he'll get the speedup, otherwise he'll get the current behavior. This is also not trivial to get right, but seems more managable than trying to write a Rectangle adapter mapping to the collection.