@timhoffm 2022-05-01
@ivirshup are you still interested in working on this (under the constraints of backward compatibility and the paths laid out in my last comment)?