@ivirshup 2022-05-04
@timhoffm, I don't think so. Many of the solutions suggested above got fairly involved, and I don't think I'm familiar enough with matplotlib internals to do this in a reasonable timeframe. 

Feel free to close, though I think the conversation here definitely has value.