@timhoffm 2022-05-04
@ivirshup thanks for the feedback. I've created issue #22976 with a concise description of the suggested solution. Therefore, I close this thread.